Observer patter redo.
