--- TopicOne/src/key_five/ItemElement.java
@@ -1,8 +1,0 @@
-package key_five;
-
-public interface ItemElement {
-
-	public float getPrice();
-
-	public String toString();
-}
