--- TopicOne/src/key_four/ItemElement.java
@@ -1,8 +1,0 @@
-package key_four;
-
-public interface ItemElement {
-
-	public float getPrice();
-
-	public String toString();
-}
