--- TopicOne/src/key_four/PaymentStrategy.java
@@ -1,10 +1,0 @@
-package key_four;
-
-import java.util.LinkedList;
-
-public interface PaymentStrategy {
-
-	public void pay(PaymentTransaction paymentTransaction);
-
-	public float calculatePrice(LinkedList<ItemElement> items);
-}
