--- TopicOne/src/key_one/CashStrategy.java
@@ -1,9 +1,0 @@
-package key_one;
-
-public class CashStrategy implements PaymentStrategy {
-
-	@Override
-	public void pay(float amount) {
-		System.out.println(amount + " paid using cash.");
-	}
-}
