--- TopicOne/src/key_one/ItemElement.java
@@ -1,6 +1,0 @@
-package key_one;
-
-public interface ItemElement {
-
-	public float getPrice();
-}
