--- TopicOne/src/key_one/PaymentStrategy.java
@@ -1,6 +1,0 @@
-package key_one;
-
-public interface PaymentStrategy {
-
-	public void pay(float amount);
-}
