+++ TopicOne/src/key_six/Mouse.java
@@ -31,6 +31,6 @@ public class Mouse implements ItemElement {
 	@Override
 	public void changePrice(float newPrice) {
 		this.price = newPrice;
-		SingletonNotificationManager.getInstance().newNotification("Mouse price changed");
+		SingletonShopping.getInstance().newNotification("Mouse price changed");
 	}
 }
