--- TopicOne/src/key_three/ItemElement.java
@@ -1,6 +1,0 @@
-package key_three;
-
-public interface ItemElement {
-
-	public float getPrice();
-}
