--- TopicOne/src/key_three/PaymentStrategy.java
@@ -1,10 +1,0 @@
-package key_three;
-
-import java.util.LinkedList;
-
-public interface PaymentStrategy {
-
-	public void pay(PaymentTransaction paymentTransaction);
-
-	public float calculatePrice(LinkedList<ItemElement> items);
-}
