--- TopicOne/src/key_three/SingletonCounter.java
@@ -1,22 +1,0 @@
-package key_three;
-
-public class SingletonCounter {
-
-	private static SingletonCounter singletonCounter = null;
-	private int counter = 0;
-
-	private SingletonCounter() {
-	}
-
-	public static SingletonCounter getInstance() {
-		if (singletonCounter == null) {
-			singletonCounter = new SingletonCounter();
-		}
-		return singletonCounter;
-	}
-
-	public int getUniqueId() {
-		counter++;
-		return counter;
-	}
-}
