--- TopicOne/src/key_two/ItemElement.java
@@ -1,6 +1,0 @@
-package key_two;
-
-public interface ItemElement {
-
-	public float getPrice();
-}
