--- TopicOne/src/key_two/PaymentStrategy.java
@@ -1,10 +1,0 @@
-package key_two;
-
-import java.util.LinkedList;
-
-public interface PaymentStrategy {
-
-	public void pay(float amount);
-
-	public float calculatePrice(LinkedList<ItemElement> items);
-}
